requirements.txt
# ── AI / RAG 
langchain>=0.2.0
langchain-community>=0.2.0
langchain-google-genai>=1.0.5
langchain-chroma>=0.1.0
google-generativeai>=0.7.0
google-genai>=0.8.0
chromadb>=0.5.0
pypdf>=4.2.0
python-dotenv>=1.0.0

# ── Web server
flask>=3.0.0
flask-cors>=4.0.0
